+++ checker/src/main/java/org/checkerframework/checker/lock/LockVisitor.java
@@ -332,7 +332,7 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
   }
 
   @Override
-  protected void commonAssignmentCheck(
+  protected boolean commonAssignmentCheck(
       AnnotatedTypeMirror varType,
       AnnotatedTypeMirror valueType,
       Tree valueTree,
@@ -346,10 +346,10 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
     // @GuardSatisfied in the "Type-checking rules" section of the Lock Checker manual chapter
     // for more details.
 
+    boolean result = true;
     if (varType.hasAnnotation(GuardSatisfied.class)) {
       if (valueType.hasAnnotation(GuardedBy.class)) {
-        checkLock(valueTree, valueType.getAnnotation(GuardedBy.class));
-        return;
+        return checkLock(valueTree, valueType.getAnnotation(GuardedBy.class));
       } else if (valueType.hasAnnotation(GuardSatisfied.class)) {
         // TODO: Find a cleaner, non-abstraction-breaking way to know whether method actual
         // parameters are being assigned to formal parameters.
@@ -366,6 +366,7 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
           if (varTypeGuardSatisfiedIndex == -1 && valueTypeGuardSatisfiedIndex == -1) {
             checker.reportError(
                 valueTree, "guardsatisfied.assignment.disallowed", varType, valueType);
+            result = false;
           }
         } else {
           // The RHS can be @GuardSatisfied with a different index when matching method
@@ -378,7 +379,7 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
           // Note: this matching of a @GS(index) to a @GS(differentIndex) is *only*
           // allowed when matching method formal parameters to actual parameters.
 
-          return;
+          return true;
         }
       } else if (!atypeFactory.getTypeHierarchy().isSubtype(valueType, varType)) {
         // Special case: replace the @GuardSatisfied primary annotation on the LHS with
@@ -387,12 +388,14 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
         AnnotatedTypeMirror varType2 = varType.deepCopy(); // TODO: Would shallowCopy be sufficient?
         varType2.replaceAnnotation(atypeFactory.GUARDEDBY);
         if (atypeFactory.getTypeHierarchy().isSubtype(valueType, varType2)) {
-          return;
+          return true;
         }
       }
     }
 
-    super.commonAssignmentCheck(varType, valueType, valueTree, errorKey, extraArgs);
+    result =
+        result && super.commonAssignmentCheck(varType, valueType, valueTree, errorKey, extraArgs);
+    return result;
   }
 
   @Override
@@ -820,8 +823,9 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
    * to be @Deterministic.
    *
    * @param lockExpressionTree the expression tree of a synchronized block
+   * @return true if the check succeeds, false if an error message was issued
    */
-  private void ensureExpressionIsEffectivelyFinal(ExpressionTree lockExpressionTree) {
+  private boolean ensureExpressionIsEffectivelyFinal(ExpressionTree lockExpressionTree) {
     // This functionality could be implemented using a visitor instead, however with this
     // design, it is easier to be certain that an error will always be issued if a tree kind is
     // not recognized.
@@ -831,6 +835,7 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
     // reporting.
     ExpressionTree tree = lockExpressionTree;
 
+    boolean result = true;
     while (true) {
       tree = TreeUtils.withoutParens(tree);
 
@@ -838,44 +843,47 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
         case MEMBER_SELECT:
           if (!isTreeSymbolEffectivelyFinalOrUnmodifiable(tree)) {
             checker.reportError(tree, "lock.expression.not.final", lockExpressionTree);
-            return;
+            return false;
           }
           tree = ((MemberSelectTree) tree).getExpression();
           break;
         case IDENTIFIER:
           if (!isTreeSymbolEffectivelyFinalOrUnmodifiable(tree)) {
             checker.reportError(tree, "lock.expression.not.final", lockExpressionTree);
+            return false;
           }
-          return;
+          return result;
         case METHOD_INVOCATION:
           Element elem = TreeUtils.elementFromUse(tree);
           if (atypeFactory.getDeclAnnotationNoAliases(elem, Deterministic.class) == null
               && atypeFactory.getDeclAnnotationNoAliases(elem, Pure.class) == null) {
             checker.reportError(tree, "lock.expression.not.final", lockExpressionTree);
-            return;
+            return false;
           }
 
           MethodInvocationTree methodInvocationTree = (MethodInvocationTree) tree;
 
           for (ExpressionTree argTree : methodInvocationTree.getArguments()) {
-            ensureExpressionIsEffectivelyFinal(argTree);
+            result = result && ensureExpressionIsEffectivelyFinal(argTree);
           }
 
           tree = methodInvocationTree.getMethodSelect();
           break;
         default:
           checker.reportError(tree, "lock.expression.possibly.not.final", lockExpressionTree);
-          return;
+          return false;
       }
     }
   }
 
-  private void ensureExpressionIsEffectivelyFinal(
+  private boolean ensureExpressionIsEffectivelyFinal(
       JavaExpression lockExpr, String expressionForErrorReporting, Tree treeForErrorReporting) {
-    if (!atypeFactory.isExpressionEffectivelyFinal(lockExpr)) {
+    boolean result = atypeFactory.isExpressionEffectivelyFinal(lockExpr);
+    if (!result) {
       checker.reportError(
           treeForErrorReporting, "lock.expression.not.final", expressionForErrorReporting);
     }
+    return result;
   }
 
   @Override
@@ -1103,8 +1111,8 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
     checkLockOfThisOrTree(tree, true, gbAnno);
   }
 
-  private void checkLock(Tree tree, AnnotationMirror gbAnno) {
-    checkLockOfThisOrTree(tree, false, gbAnno);
+  private boolean checkLock(Tree tree, AnnotationMirror gbAnno) {
+    return checkLockOfThisOrTree(tree, false, gbAnno);
   }
 
   /**
@@ -1113,39 +1121,47 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
    * @param tree a tree whose lock to check
    * @param implicitThis true if checking the lock of the implicit {@code this}
    * @param gbAnno a @GuardedBy annotation
+   * @return true if the check succeeds, false if an error message was issued
    */
-  private void checkLockOfThisOrTree(Tree tree, boolean implicitThis, AnnotationMirror gbAnno) {
+  private boolean checkLockOfThisOrTree(Tree tree, boolean implicitThis, AnnotationMirror gbAnno) {
     if (gbAnno == null) {
       throw new TypeSystemError("LockVisitor.checkLock: gbAnno cannot be null");
     }
     if (atypeFactory.areSameByClass(gbAnno, GuardedByUnknown.class)
         || atypeFactory.areSameByClass(gbAnno, GuardedByBottom.class)) {
       checker.reportError(tree, "lock.not.held", "unknown lock " + gbAnno);
-      return;
+      return false;
     } else if (atypeFactory.areSameByClass(gbAnno, GuardSatisfied.class)) {
-      return;
+      return true;
     }
 
     List<LockExpression> expressions = getLockExpressions(implicitThis, gbAnno, tree);
     if (expressions.isEmpty()) {
-      return;
+      return true;
     }
 
+    boolean result = true;
     LockStore store = atypeFactory.getStoreBefore(tree);
     for (LockExpression expression : expressions) {
       if (expression.error != null) {
         checker.reportError(tree, "expression.unparsable", expression.error.toString());
+        result = false;
       } else if (expression.lockExpression == null) {
         checker.reportError(tree, "expression.unparsable", expression.expressionString);
+        result = false;
       } else if (!isLockHeld(expression.lockExpression, store)) {
         checker.reportError(tree, "lock.not.held", expression.lockExpression.toString());
+        result = false;
       }
 
       if (expression.error != null && expression.lockExpression != null) {
-        ensureExpressionIsEffectivelyFinal(
-            expression.lockExpression, expression.expressionString, tree);
+        result =
+            result
+                && ensureExpressionIsEffectivelyFinal(
+                    expression.lockExpression, expression.expressionString, tree);
       }
     }
+    return result;
   }
 
   private boolean isLockHeld(JavaExpression lockExpr, LockStore store) {